moving getReorderIndex to reparent-strategy-helpers to fix dependency cruiser
