+++ editor/src/components/canvas/canvas-strategies/flex-reparent-to-flex-strategy.tsx
@@ -1,27 +1,9 @@
 import { MetadataUtils } from '../../../core/model/element-metadata-utils'
-import * as EP from '../../../core/shared/element-path'
-import { offsetPoint } from '../../../core/shared/math-utils'
-import { CSSCursor } from '../canvas-types'
-import { CanvasCommand } from '../commands/commands'
-import { reorderElement } from '../commands/reorder-element-command'
-import { reparentElement } from '../commands/reparent-element-command'
-import { setCursorCommand } from '../commands/set-cursor-command'
-import { setElementsToRerenderCommand } from '../commands/set-elements-to-rerender-command'
-import { updateHighlightedViews } from '../commands/update-highlighted-views-command'
-import { updateSelectedViews } from '../commands/update-selected-views-command'
 import { ParentBounds } from '../controls/parent-bounds'
 import { ParentOutlines } from '../controls/parent-outlines'
 import { DragOutlineControl } from '../controls/select-mode/drag-outline-control'
-import { CanvasStrategy, emptyStrategyApplicationResult } from './canvas-strategy-types'
-import { getReorderIndex } from './flex-reorder-strategy'
-import { ifAllowedToReparent } from './reparent-helpers'
-import {
-  applyFlexReparent,
-  findReparentStrategy,
-  getReparentTargetForFlexElement,
-} from './reparent-strategy-helpers'
-import { getReparentCommands } from './reparent-utils'
-import { getDragTargets } from './shared-absolute-move-strategy-helpers'
+import { CanvasStrategy } from './canvas-strategy-types'
+import { applyFlexReparent, findReparentStrategy } from './reparent-strategy-helpers'
 
 export const flexReparentToFlexStrategy: CanvasStrategy = {
   id: 'FLEX_REPARENT_TO_FLEX',
